refactor: Update Fixtures page to use Material-UI Table component
This commit updates the Fixtures page to use the Material-UI Table component for displaying the player table. The existing HTML table has been replaced with the Table, TableContainer, TableHead, TableCell, and TableRow components from the @mui/material package. This change improves the overall styling and functionality of the player table.