feat: Add react-icons library for additional icons
This commit adds the react-icons library to the project dependencies. The react-icons library provides a wide range of icons that can be easily used in React applications. By incorporating react-icons, we can enhance the visual appeal and user experience of the application by adding a variety of icons to different components and pages.